Fixed several issues, including TREPrivateData flush, TRE non-default handling, and several AIS patches that hadnt made it out yet.
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@352 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c/nitf/include/nitf/BandSource.h
@@ -69,7 +69,7 @@ typedef nitf_DataSource    nitf_BandSource;
  *          out how many actual bytes to skip)
  */
 NITFAPI(nitf_BandSource *) nitf_MemorySource_construct(char *data,
-        size_t size,
+        off_t size,
         off_t start,
         int numBytesPerPixel,
         int pixelSkip,
